Common: Move NonCopyable to its own header
