+++ Source/Core/AudioCommon/WaveFile.h
@@ -17,6 +17,7 @@
 #include <string>
 #include "Common/CommonTypes.h"
 #include "Common/FileUtil.h"
+#include "Common/NonCopyable.h"
 
 class WaveFileWriter : NonCopyable
 {
