+++ Source/Core/Common/CodeBlock.h
@@ -6,6 +6,7 @@
 
 #include "Common/Common.h"
 #include "Common/MemoryUtil.h"
+#include "Common/NonCopyable.h"
 
 // Everything that needs to generate code should inherit from this.
 // You get memory management for free, plus, you can use all emitter functions without
