+++ Source/Core/Common/FileUtil.h
@@ -11,6 +11,7 @@
 #include <vector>
 
 #include "Common/Common.h"
+#include "Common/NonCopyable.h"
 
 #ifdef _WIN32
 #include "Common/StringUtil.h"
