+++ Source/Core/Common/Logging/LogManager.h
@@ -11,6 +11,7 @@
 #include <string>
 
 #include "Common/Common.h"
+#include "Common/NonCopyable.h"
 
 #define MAX_MESSAGES 8000
 #define MAX_MSGLEN  1024
