+++ Source/Core/Core/DSP/DSPCaptureLogger.h
@@ -7,7 +7,8 @@
 #include <memory>
 #include <string>
 
-#include "Common/Common.h"
+#include "Common/CommonTypes.h"
+#include "Common/NonCopyable.h"
 
 class PCAP;
