+++ Source/Core/Core/HW/GCMemcard.h
@@ -10,6 +10,7 @@
 #include "Common/CommonPaths.h"
 #include "Common/CommonTypes.h"
 #include "Common/NandPaths.h"
+#include "Common/NonCopyable.h"
 #include "Common/StringUtil.h"
 
 #include "Core/HW/EXI_DeviceIPL.h"
