+++ Source/Core/Core/IPC_HLE/WII_Socket.h
@@ -47,6 +47,8 @@ typedef struct pollfd pollfd_t;
 #include <string>
 #include <unordered_map>
 
+#include "Common/CommonTypes.h"
+#include "Common/NonCopyable.h"
 #include "Core/IPC_HLE/WII_IPC_HLE.h"
 #include "Core/IPC_HLE/WII_IPC_HLE_Device_net.h"
 #include "Core/IPC_HLE/WII_IPC_HLE_Device_net_ssl.h"
